Changed getters to match rust convention as pointed out in issue https://github.com/tomaka/cpal/issues/91
