+++ src/alsa/enumerate.rs
@@ -112,8 +112,3 @@ impl Iterator for EndpointsIterator {
         }
     }
 }
-
-#[inline]
-pub fn get_default_endpoint() -> Option<Endpoint> {
-    Some(Endpoint("default".to_owned()))
-}
