+++ src/coreaudio/mod.rs
@@ -23,14 +23,17 @@ use self::coreaudio::audio_unit::render_callback::{self, data};
 mod enumerate;
 
 pub use self::enumerate::{EndpointsIterator,
-                          SupportedFormatsIterator,
-                          get_default_endpoint};
+                          SupportedFormatsIterator};
 
 #[derive(Clone, PartialEq, Eq)]
 pub struct Endpoint;
 
 impl Endpoint {
-    pub fn get_supported_formats_list(&self)
+    pub fn default_endpoint() -> Option<Endpoint> {
+        Some(Endpoint)
+    }
+
+    pub fn supported_formats(&self)
             -> Result<SupportedFormatsIterator, FormatsEnumerationError>
     {
         Ok(vec!(Format {
@@ -40,7 +43,7 @@ impl Endpoint {
         }).into_iter())
     }
 
-    pub fn get_name(&self) -> String {
+    pub fn name(&self) -> String {
         "Default AudioUnit Endpoint".to_string()
     }
 }
@@ -60,7 +63,7 @@ pub struct Buffer<T> {
 
 impl<T> Buffer<T> where T: Sample {
     #[inline]
-    pub fn get_buffer(&mut self) -> &mut [T] {
+    pub fn buffer(&mut self) -> &mut [T] {
         &mut self.buffer[..]
     }
