+++ src/null/mod.rs
@@ -31,24 +31,24 @@ impl Iterator for EndpointsIterator {
     }
 }
 
-#[inline]
-pub fn get_default_endpoint() -> Option<Endpoint> {
-    None
-}
-
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Endpoint;
 
 impl Endpoint {
     #[inline]
-    pub fn get_supported_formats_list(&self)
+    pub fn default_endpoint() -> Option<Endpoint> {
+        None
+    }
+
+    #[inline]
+    pub fn supported_formats(&self)
             -> Result<SupportedFormatsIterator, FormatsEnumerationError>
     {
         unreachable!()
     }
 
     #[inline]
-    pub fn get_name(&self) -> String {
+    pub fn name(&self) -> String {
         "null".to_owned()
     }
 }
@@ -100,7 +100,7 @@ pub struct Buffer<T> {
 
 impl<T> Buffer<T> {
     #[inline]
-    pub fn get_buffer(&mut self) -> &mut [T] {
+    pub fn buffer(&mut self) -> &mut [T] {
         unreachable!()
     }
