+++ src/samples_formats.rs
@@ -14,7 +14,7 @@ pub enum SampleFormat {
 impl SampleFormat {
     /// Returns the size in bytes of a sample of this format.
     #[inline]
-    pub fn get_sample_size(&self) -> usize {
+    pub fn sample_size(&self) -> usize {
         match self {
             &SampleFormat::I16 => mem::size_of::<i16>(),
             &SampleFormat::U16 => mem::size_of::<u16>(),
@@ -26,26 +26,26 @@ impl SampleFormat {
 /// Trait for containers that contain PCM data.
 pub unsafe trait Sample: Copy + Clone {
     /// Returns the `SampleFormat` corresponding to this data type.
-    fn get_format() -> SampleFormat;
+    fn format() -> SampleFormat;
 }
 
 unsafe impl Sample for u16 {
     #[inline]
-    fn get_format() -> SampleFormat {
+    fn format() -> SampleFormat {
         SampleFormat::U16
     }
 }
 
 unsafe impl Sample for i16 {
     #[inline]
-    fn get_format() -> SampleFormat {
+    fn format() -> SampleFormat {
         SampleFormat::I16
     }
 }
 
 unsafe impl Sample for f32 {
     #[inline]
-    fn get_format() -> SampleFormat {
+    fn format() -> SampleFormat {
         SampleFormat::F32
     }
 }
